Improved fetching GitHub issues.
We're using a single rest call to acquire the issues.

+++ buildSrc/src/main/groovy/org/mockito/release/notes/improvements/Improvement.java
@@ -6,12 +6,12 @@ import java.util.Set;
 
 class Improvement {
 
-    private final int id;
+    private final long id;
     private final String title;
     private final String url;
     private final Collection<String> labels;
 
-    public Improvement(int id, String title, String url, Collection<String> labels) {
+    public Improvement(long id, String title, String url, Collection<String> labels) {
         this.id = id;
         this.title = title;
         this.url = url;
